EZP-24361: Pass field instead of content + fieldDefIdentifier to field renderer.
This makes it consistent with `renderFieldDefinitionView()` and `renderFieldDefinitionEdit()` methods.
Parameters like content, contentInfo and versionInfo are now being passed by the Twig extension directly.
