+++ eZ/Bundle/EzPublishCoreBundle/Resources/config/templating.yml
@@ -137,7 +137,11 @@ services:
 
     ezpublish.twig.extension.field_rendering:
         class: %ezpublish.twig.extension.field_rendering.class%
-        arguments: [@ezpublish.templating.field_block_renderer]
+        arguments:
+            - @ezpublish.templating.field_block_renderer
+            - @ezpublish.api.service.content_type
+            - @ezpublish.fieldType.parameterProviderRegistry
+            - @ezpublish.translation_helper
         tags:
             - { name: twig.extension }
