+++ eZ/Publish/Core/MVC/Symfony/Templating/FieldBlockRendererInterface.php
@@ -10,6 +10,7 @@
 namespace eZ\Publish\Core\MVC\Symfony\Templating;
 
 use eZ\Publish\API\Repository\Values\Content\Content;
+use eZ\Publish\API\Repository\Values\Content\Field;
 use eZ\Publish\API\Repository\Values\ContentType\FieldDefinition;
 
 /**
@@ -21,26 +22,26 @@ interface FieldBlockRendererInterface
     /**
      * Renders the HTML view markup for a given field.
      *
-     * @param \eZ\Publish\API\Repository\Values\Content\Content $content
-     * @param string $fieldIdentifier Identifier for the field we want to view
+     * @param \eZ\Publish\API\Repository\Values\Content\Field $field
+     * @param string $fieldTypeIdentifier FieldType identifier for $field
      * @param array $params An array of parameters to pass to the field view
      *
      * @throws \eZ\Publish\Core\Base\Exceptions\InvalidArgumentException If no field with provided $fieldIdentifier can be found in $content.
      * @return string
      */
-    public function renderContentFieldView( Content $content, $fieldIdentifier, array $params = [] );
+    public function renderContentFieldView( Field $field, $fieldTypeIdentifier, array $params = [] );
 
     /**
      * Renders the HTML edit markup for a given field.
      *
-     * @param \eZ\Publish\API\Repository\Values\Content\Content $content
-     * @param string $fieldIdentifier Identifier for the field we want to edit
+     * @param \eZ\Publish\API\Repository\Values\Content\Field $field
+     * @param string $fieldTypeIdentifier FieldType identifier for $field
      * @param array $params An array of parameters to pass to the field edit view
      *
      * @throws \eZ\Publish\Core\Base\Exceptions\InvalidArgumentException If no field with provided $fieldIdentifier can be found in $content.
      * @return string
      */
-    public function renderContentFieldEdit( Content $content, $fieldIdentifier, array $params = [] );
+    public function renderContentFieldEdit( Field $field, $fieldTypeIdentifier, array $params = [] );
 
     /**
      * Renders the HTML view markup for the given field definition.
